requirements.txt
# Core dependencies (no Abaqus required for unit tests)
pyyaml>=6.0
jsonschema>=4.0

# Server (FastAPI for HTTP, MCP for Model Context Protocol)
fastapi>=0.100
uvicorn>=0.20
mcp>=1.0

# LLM backends (optional, install as needed)
# anthropic>=0.36
# openai>=1.0

# Testing
pytest>=7.0
pytest-cov>=4.0

# Code quality
ruff>=0.1
